Changed npe to a double, in case there's really no photostatistics

+++ SimCalorimetry/CaloSimAlgos/src/CaloHitResponse.cc
@@ -97,10 +97,10 @@ CaloSamples CaloHitResponse::makeAnalogSignal(const PCaloHit & inputHit) const {
 double CaloHitResponse::analogSignalAmplitude(const PCaloHit & hit, const CaloSimParameters & parameters) const {
   // OK, the "energy" in the hit could be a real energy, deposited energy,
   // or pe count.  This factor converts to photoelectrons
-  int npe = (int)(hit.energy() * parameters.simHitToPhotoelectrons());
+  double npe = hit.energy() * parameters.simHitToPhotoelectrons();
   // do we need to doPoisson statistics for the photoelectrons?
   if(parameters.doPhotostatistics()) {
-    npe = RandPoisson::shoot(npe);
+    npe = RandPoisson::shoot(static_cast<int>(npe));
   }
   return npe;
 }
